+++ src/review/semantic.ts
@@ -18,16 +18,23 @@ import type { DebateRunnerOptions } from "../debate";
 import { getSafeLogger } from "../logger";
 import { callOp } from "../operations/call";
 import { semanticReviewOp } from "../operations/semantic-review";
-import type { ReviewFinding } from "../plugins/types";
 import type { UserStory } from "../prd";
 import { ReviewPromptBuilder } from "../prompts";
-import { createRuntime } from "../runtime";
 import { formatSessionName } from "../session/naming";
 import { resolveReviewExcludePatterns, resolveTestFilePatterns } from "../test-runners";
-import { tryParseLLMJson } from "../utils/llm-json";
 import type { NaxIgnoreIndex } from "../utils/path-filters";
 import { DIFF_CAP_BYTES, collectDiff, collectDiffStat, resolveEffectiveRef, truncateDiff } from "./diff-utils";
 import { writeReviewAudit } from "./review-audit";
+import { runSemanticDebate } from "./semantic-debate";
+import {
+  type LLMFinding,
+  type LLMResponse,
+  formatFindings,
+  isBlockingSeverity,
+  parseLLMResponse,
+  sanitizeRefModeFindings,
+  toReviewFindings,
+} from "./semantic-helpers";
 import { looksLikeTruncatedJson } from "./truncation";
 import type { ReviewCheckResult, SemanticReviewConfig, SemanticStory } from "./types";
 
@@ -39,139 +46,6 @@ export const _semanticDeps = {
   createDebateRunner: (opts: DebateRunnerOptions): DebateRunner => new DebateRunner(opts),
   writeReviewAudit,
 };
-
-interface LLMFinding {
-  severity: string;
-  file: string;
-  line: number;
-  issue: string;
-  suggestion: string;
-  acId?: string;
-  verifiedBy?: {
-    command?: string;
-    file: string;
-    line?: number;
-    observed: string;
-  };
-}
-
-interface LLMResponse {
-  passed: boolean;
-  findings: LLMFinding[];
-}
-
-/**
- * Validate parsed JSON matches the expected LLM response shape.
- */
-function validateLLMShape(parsed: unknown): LLMResponse | null {
-  if (typeof parsed !== "object" || parsed === null) return null;
-  const obj = parsed as Record<string, unknown>;
-  if (typeof obj.passed !== "boolean") return null;
-  if (!Array.isArray(obj.findings)) return null;
-  return { passed: obj.passed, findings: obj.findings as LLMFinding[] };
-}
-
-/**
- * Parse and validate LLM JSON response using multi-tier extraction.
- * Returns null only when all tiers fail or shape validation fails.
- */
-function parseLLMResponse(raw: string): LLMResponse | null {
-  try {
-    return validateLLMShape(tryParseLLMJson(raw));
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Format findings into readable text output.
- */
-function formatFindings(findings: LLMFinding[]): string {
-  return findings
-    .map((f) => `[${f.severity}] ${f.file}:${f.line} — ${f.issue}\n  Suggestion: ${f.suggestion}`)
-    .join("\n");
-}
-
-/** Normalize LLM severity values to ReviewFinding severity union. */
-function normalizeSeverity(sev: string): ReviewFinding["severity"] {
-  if (sev === "warn") return "warning";
-  if (sev === "unverifiable") return "info";
-  if (sev === "critical" || sev === "error" || sev === "warning" || sev === "info" || sev === "low") return sev;
-  return "info";
-}
-
-/**
- * Severity rank for threshold comparison.
- * "unverifiable" is treated as info (non-blocking by default).
- */
-const SEVERITY_RANK: Record<string, number> = {
-  info: 0,
-  unverifiable: 0,
-  low: 1,
-  warning: 1,
-  error: 2,
-  critical: 3,
-};
-
-const UNVERIFIED_FINDING_PATTERNS = [
-  "cannot verify",
-  "can't verify",
-  "from diff alone",
-  "missing from diff",
-  "not found in diff",
-  "not present in diff",
-  "does not appear in diff",
-] as const;
-
-/**
- * Check whether a normalized finding severity meets or exceeds the blocking threshold.
- * threshold defaults to "error" — only error/critical block unless configured stricter.
- */
-function isBlockingSeverity(sev: string, threshold: "error" | "warning" | "info" = "error"): boolean {
-  return (SEVERITY_RANK[sev] ?? 0) >= (SEVERITY_RANK[threshold] ?? 2);
-}
-
-/** Ref-mode semantic errors must prove they were verified against current files. */
-function sanitizeRefModeFindings(findings: LLMFinding[], diffMode: SemanticReviewConfig["diffMode"]): LLMFinding[] {
-  if (diffMode !== "ref") return findings;
-  return findings.map((finding) =>
-    needsDowngradeForMissingEvidence(finding) ? downgradeToUnverifiable(finding) : finding,
-  );
-}
-
-function needsDowngradeForMissingEvidence(finding: LLMFinding): boolean {
-  if ((SEVERITY_RANK[finding.severity] ?? 0) < SEVERITY_RANK.error) return false;
-  return mentionsUnverifiedSource(finding) || !hasVerifiedEvidence(finding);
-}
-
-function mentionsUnverifiedSource(finding: LLMFinding): boolean {
-  const text = `${finding.issue} ${finding.suggestion}`.toLowerCase();
-  return UNVERIFIED_FINDING_PATTERNS.some((pattern) => text.includes(pattern));
-}
-
-function hasVerifiedEvidence(finding: LLMFinding): boolean {
-  const evidence = finding.verifiedBy;
-  return !!evidence?.file?.trim() && !!evidence.observed?.trim();
-}
-
-function downgradeToUnverifiable(finding: LLMFinding): LLMFinding {
-  return {
-    ...finding,
-    severity: "unverifiable",
-  };
-}
-
-/** Convert LLMFinding[] to ReviewFinding[] with semantic-review metadata. */
-function toReviewFindings(findings: LLMFinding[]): ReviewFinding[] {
-  return findings.map((f) => ({
-    ruleId: "semantic",
-    severity: normalizeSeverity(f.severity),
-    file: f.file,
-    line: f.line,
-    message: f.issue,
-    source: "semantic-review",
-  }));
-}
 
 /**
  * Run a semantic review using an LLM against the story diff.
@@ -306,194 +180,24 @@ export async function runSemanticReview(
   // Debate path: when debate is enabled for review stage, use DebateRunner instead of agent.complete()
   const reviewDebateEnabled = naxConfig?.debate?.enabled && naxConfig?.debate?.stages?.review?.enabled;
   if (reviewDebateEnabled) {
-    const semanticEffectiveConfig = naxConfig ?? DEFAULT_CONFIG;
-    const createdRuntime = runtime
-      ? undefined
-      : createRuntime(semanticEffectiveConfig, workdir, { agentManager, featureName: featureName ?? "_standalone" });
-    const debateRuntime = runtime ?? createdRuntime;
-    if (!debateRuntime) {
-      throw new Error("[review] Debate runtime unavailable");
-    }
-    try {
-      // Safe: reviewDebateEnabled guard confirms naxConfig.debate.stages.review is defined
-      const configuredStageConfig = naxConfig?.debate?.stages.review as import("../debate").DebateStageConfig;
-      const reviewStageConfig =
-        configuredStageConfig.sessionMode === "one-shot" && (configuredStageConfig.mode ?? "panel") === "panel"
-          ? configuredStageConfig
-          : {
-              ...configuredStageConfig,
-              // Review debate currently supports panel one-shot only.
-              sessionMode: "one-shot" as const,
-              mode: "panel" as const,
-            };
-      if (reviewStageConfig !== configuredStageConfig) {
-        logger?.warn("review", "Review debate requires sessionMode=one-shot and mode=panel — forcing safe defaults", {
-          storyId: story.id,
-          configuredSessionMode: configuredStageConfig.sessionMode,
-          configuredMode: configuredStageConfig.mode ?? "panel",
-        });
-      }
-      const isReReview = resolverSession !== undefined && resolverSession.history.length > 0;
-      const semanticAgentName =
-        agentManager && typeof (agentManager as IAgentManager).getDefault === "function"
-          ? (agentManager as IAgentManager).getDefault()
-          : "claude";
-      const semanticCallCtx: import("../operations/types").CallContext = {
-        runtime: debateRuntime,
-        packageView: debateRuntime.packages.resolve(workdir),
-        packageDir: workdir,
-        agentName: semanticAgentName,
-        storyId: story.id,
-        featureName,
-      };
-      const debateRunner = _semanticDeps.createDebateRunner({
-        ctx: semanticCallCtx,
-        stage: "review",
-        stageConfig: reviewStageConfig,
-        config: semanticEffectiveConfig,
-        workdir,
-        featureName: featureName,
-        timeoutSeconds: naxConfig?.execution?.sessionTimeoutSeconds,
-        reviewerSession: resolverSession,
-        resolverContextInput: resolverSession
-          ? {
-              diffMode,
-              ...(diffMode === "ref" ? { storyGitRef: effectiveRef, stat } : { diff }),
-              story: { id: story.id, title: story.title, acceptanceCriteria: story.acceptanceCriteria },
-              semanticConfig,
-              resolverType: reviewStageConfig.resolver.type,
-              isReReview,
-            }
-          : undefined,
-      });
-      // Track history length before to detect if the session was actually used by the resolver
-      const historyLenBefore = resolverSession?.history.length ?? 0;
-      const debateResult = await debateRunner.run(prompt);
-      const debateCost = debateResult.totalCostUsd ?? 0;
-
-      // When the ReviewerSession was used by the resolver (history grew), use its tool-verified
-      // verdict via getVerdict() instead of re-deriving from raw proposals.
-      const sessionUsed = resolverSession && resolverSession.history.length > historyLenBefore;
-      if (sessionUsed) {
-        const durationMs = Date.now() - startTime;
-        try {
-          const verdict = resolverSession.getVerdict();
-          const findings = verdict.findings ?? [];
-          if (!verdict.passed && findings.length > 0) {
-            logger?.warn("review", `Semantic review failed (debate+dialogue): ${findings.length} findings`, {
-              storyId: story.id,
-              durationMs,
-            });
-            return {
-              check: "semantic",
-              success: false,
-              command: "",
-              exitCode: 1,
-              output: `Semantic review failed:\n\n${findings.map((f) => `${f.ruleId}: ${f.message}`).join("\n")}`,
-              durationMs,
-              findings,
-              cost: debateCost,
-            };
-          }
-          const label = verdict.passed
-            ? "Semantic review passed (debate+dialogue)"
-            : "Semantic review passed (debate+dialogue, all findings non-blocking)";
-          logger?.info("review", label, { storyId: story.id, durationMs });
-          return {
-            check: "semantic",
-            success: true,
-            command: "",
-            exitCode: 0,
-            output: label,
-            durationMs,
-            cost: debateCost,
-          };
-        } catch {
-          // getVerdict() threw (e.g. session destroyed) — fall through to stateless path
-          logger?.warn("review", "getVerdict() failed after debate+dialogue — falling back to stateless verdict", {
-            storyId: story.id,
-          });
-        }
-      }
-
-      // Stateless fallback: re-derive verdict from proposals (existing behavior)
-      const resolverPassed = debateResult.outcome === "passed";
-      const allFindings: LLMFinding[] = [];
-      for (const p of debateResult.proposals) {
-        const parsed = parseLLMResponse(p.output);
-        if (parsed) {
-          allFindings.push(...parsed.findings);
-        }
-      }
-
-      // Deduplicate findings by AC id (primary) or file:line (fallback)
-      const seen = new Set<string>();
-      const deduped: LLMFinding[] = [];
-      for (const f of allFindings) {
-        const key = f.acId ?? `${f.file}:${f.line}`;
-        if (!seen.has(key)) {
-          seen.add(key);
-          deduped.push(f);
-        }
-      }
-
-      // Split debate findings by blocking threshold
-      const debateFindings = sanitizeRefModeFindings(deduped, diffMode);
-      const debateThreshold = blockingThreshold ?? "error";
-      const debateBlocking = debateFindings.filter((f) => isBlockingSeverity(f.severity, debateThreshold));
-      const debateAdvisory = debateFindings.filter((f) => !isBlockingSeverity(f.severity, debateThreshold));
-
-      const durationMs = Date.now() - startTime;
-      if (!resolverPassed) {
-        if (debateBlocking.length > 0) {
-          logger?.warn("review", `Semantic review failed (debate): ${debateBlocking.length} blocking findings`, {
-            storyId: story.id,
-            durationMs,
-          });
-          return {
-            check: "semantic",
-            success: false,
-            command: "",
-            exitCode: 1,
-            output: `Semantic review failed:\n\n${formatFindings(debateBlocking)}`,
-            durationMs,
-            findings: toReviewFindings(debateBlocking),
-            advisoryFindings: debateAdvisory.length > 0 ? toReviewFindings(debateAdvisory) : undefined,
-            cost: debateCost,
-          };
-        }
-        // All findings were advisory — override to pass
-        logger?.info("review", "Semantic review passed (debate, all findings below blocking threshold)", {
-          storyId: story.id,
-          durationMs,
-        });
-        return {
-          check: "semantic",
-          success: true,
-          command: "",
-          exitCode: 0,
-          output: "Semantic review passed (debate, all findings were advisory — below blocking threshold)",
-          durationMs,
-          advisoryFindings: debateAdvisory.length > 0 ? toReviewFindings(debateAdvisory) : undefined,
-          cost: debateCost,
-        };
-      }
-      logger?.info("review", "Semantic review passed (debate)", { storyId: story.id, durationMs });
-      return {
-        check: "semantic",
-        success: true,
-        command: "",
-        exitCode: 0,
-        output: "Semantic review passed",
-        durationMs,
-        advisoryFindings: debateAdvisory.length > 0 ? toReviewFindings(debateAdvisory) : undefined,
-        cost: debateCost,
-      };
-    } finally {
-      if (createdRuntime) {
-        await createdRuntime.close().catch(() => {});
-      }
-    }
+    return runSemanticDebate({
+      naxConfig,
+      runtime,
+      workdir,
+      agentManager,
+      featureName,
+      story,
+      resolverSession,
+      diffMode,
+      diff,
+      stat,
+      semanticConfig,
+      effectiveRef,
+      startTime,
+      prompt,
+      blockingThreshold,
+      createDebateRunner: _semanticDeps.createDebateRunner,
+    });
   }
 
   // ADR-019 Pattern A: when a NaxRuntime is available, dispatch via callOp so the
